More Tests and Fixes
- Fixed delay issues with sound (decompressing mp3 files was introducing an unacceptable gap between sound sources, used uncompressed WAV to resolve)
- Fixed recoil damage and health absorption allowing values of 0 when the minimum should be 1
- Fixed TeamBattleModifier constructor initializing wild pokemon prefix to the wrong value
- Currently trying to devise a way to easily manage execution of game events without having 6 billion unnecessary components. Leaning toward a ScriptableObject implementation, although this has the drawback of not allowing coroutine origination from the event object. Furthermore, guaranteeing that the necessary methods will be available at compile time is tricky since interfaces are not serializable in Unity, go figure.
Perhaps it would be better to have InteractEvent components submit themselves to a universal event handler that does all the work? How to handle events where subsequent events rely on the outcome of a previous one? Ordinarily a method could return a value that's used by the next, but to utilize the benefits of waiting via coroutine, the method must return IEnumerator, so return values can't be used to manage event outcomes.
I sure hope nobody is actually reading this.